+++ src/Services/Booking/WaitlistClient/Result/WaitlistClientItemResult.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Veronica Akhmetova <264936994+fatestr1ngs@users.noreply.github.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Booking\WaitlistClient\Result;
+
+use Bitrix24\SDK\Core\Result\AbstractAnnotatedItem;
+
+/**
+ * @property-read int|null $id
+ * @property-read array<string, string>|null $type
+ */
+class WaitlistClientItemResult extends AbstractAnnotatedItem
+{
+}